Playing Cards
When card is drawn the deck script sets the card's id on the dragg
script. When the card is played passes in its id to the playedCard() on
the deck script which removes it from the hand list.

I changed how the start hand is drawn for testing purposes but we can
change that back later (currently you start with 0 cards and 1st
spacebar press draws start hand).